requirements.txt
flask==2.3.3
flask-cors==4.0.0
asyncpg==0.28.0
aioredis==2.0.1
aiohttp==3.8.5
psutil==5.9.5
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
redis==4.6.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
pandas==2.0.3
sqlalchemy==2.0.20
alembic==1.12.0

